pr-style: fix with codefactor style
