+++ src/Arcus.WebApi.Tests.Integration/Logging/Fixture/InMemoryApplicationInsightsTelemetryConverter.cs
@@ -2,11 +2,8 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Arcus.Observability.Telemetry.Serilog.Sinks.ApplicationInsights.Converters;
 using Microsoft.ApplicationInsights.Channel;
-using Microsoft.ApplicationInsights.Extensibility;
 using Serilog.Events;
 using Serilog.Sinks.ApplicationInsights.Sinks.ApplicationInsights.TelemetryConverters;
 
@@ -37,41 +34,5 @@ namespace Arcus.WebApi.Tests.Integration.Logging.Fixture
 
             return Enumerable.Empty<ITelemetry>();
         }
-    }
-
-    public class InMemoryTelemetryProcessor : ITelemetryProcessor
-    {
-        private readonly ConcurrentStack<ITelemetry> _telemetries = new ConcurrentStack<ITelemetry>();
-
-        public ITelemetry[] Telemetries => _telemetries.ToArray();
-
-        public void Process(ITelemetry item)
-        {
-            _telemetries.Push(item);
-        }
-    }
-
-    public class InMemoryTelemetryChannel : ITelemetryChannel
-    {
-        private readonly ConcurrentStack<ITelemetry> _telemetries = new ConcurrentStack<ITelemetry>();
-
-        public ITelemetry[] Telemetries => _telemetries.ToArray(); 
-
-        public void Dispose()
-        {
-            
-        }
-
-        public void Send(ITelemetry item)
-        {
-            _telemetries.Push(item);
-        }
-
-        public void Flush()
-        {
-        }
-
-        public bool? DeveloperMode { get; set; }
-        public string EndpointAddress { get; set; }
     }
 }
